+++ helicopter-pilot-demo/src/components/charts/barChart.tsx
@@ -1,97 +1,77 @@
+import "./styles.css";
 import React from "react";
-import {Column, DualAxes} from "@ant-design/charts";
+import {
+  ComposedChart,
+  Bar,
+  XAxis,
+  YAxis,
+  CartesianGrid,
+  Tooltip,
+  Legend,
+  Line,
+  Label
+} from "recharts";
 
+const data = [
+    {
+        "datetime": "2021-01",
+        "max": 18,
+        "Shell": 7,
+        "Equinor": 7,
+    },
+    {
+        "datetime": "2021-02",
+        "max": 18,
+        "Shell": 7,
+        "Equinor": 9,
+    },
+    {
+        "datetime": "2021-03",
+        "max": 18,
+        "Shell": 7,
+        "Equinor": 4,
+    },
+    {
+        "datetime": "2021-04",
+        "max": 18,
+        "Shell": 5,
+        "Equinor": 10,
+    },
+    {
+        "datetime": "2021-05",
+        "max": 18,
+        "Shell": 7,
+        "Equinor": 2,
+    },
+    {
+        "datetime": "2021-06",
+        "max": 18,
+        "Shell": 3,
+        "Equinor": 7,
+    },
+];
 
 export function BarChart() {
-    var data = [
-        {
-            "datetime": "2021-01",
-            "max": 18,
-            "value": 7,
-            "company": "Shell",
-        },
-        {
-            "datetime": "2021-02",
-            "max": 18,
-            "value": 5,
-            "company": "Shell",
-        },
-        {
-            "datetime": "2021-03",
-            "max": 18,
-            "value": 5,
-            "company": "Shell",
-        },
-        {
-            "datetime": "2021-04",
-            "max": 18,
-            "value": 9,
-            "company": "Shell",
-        },
-        {
-            "datetime": "2021-01",
-            "max": 18,
-            "value": 5,
-            "company": "Equinor",
-        },
-        {
-            "datetime": "2021-02",
-            "max": 18,
-            "value": 9,
-            "company": "Equinor",
-        },
-        {
-            "datetime": "2021-03",
-            "max": 18,
-            "value": 3,
-            "company": "Equinor",
-        },
-        {
-            "datetime": "2021-04",
-            "max": 18,
-            "value": 7,
-            "company": "Equinor",
-        },
-    ];
-
-    var config = {
-        data: [data, data],
-
-        xField: "datetime",
-        yField: ["value", "max"],
-        yAxis: {
-            max: {
-                min: 0,
-                max: 20,
-                top: true,
-                autoHide: true
-            },
-            num: {
-                min: 0,
-                max: 20,
-                top: true,
-                        }
-        },
-        geometryOptions: [
-            { geometry: 'column',
-            seriesField: "company",
-            color: ['#f5c82a', '#DC143C'],
-            isStack: true,
-        },
-            {
-              geometry: 'line',
-            },
-          ],
-        interactions: [
-            {
-              type: 'active-region',
-              enable: false,
-            },
-          ]
-            };
-    return (
-        <>
-            < DualAxes{...config} width={1000} />
-        </>
-    );
+  return (
+    <ComposedChart
+      width={1000}
+      height={500}
+      data={data}
+      margin={{
+        top: 20,
+        right: 30,
+        left: 100,
+        bottom: 5
+      }}
+    >
+      <CartesianGrid strokeDasharray="3 3" />
+      <XAxis dataKey="datetime" />
+      <YAxis />
+      <Tooltip />
+      <Legend />
+      <Bar dataKey="Shell" stackId="a" fill="#f5c82a" />
+      <Bar dataKey="Equinor" stackId="a" fill="#DC143C"/>
+      <Line type="linear" dataKey="max" stroke="#b0b0b0"/>
+    </ComposedChart>
+  );
 }
